Updated branch to be in conformance with Sprint-3
Issues regarding import of module typeDefs wer resolved in this branch
after changes were made to the files in typing, specifically in the
config folder which did not exist in Main branch. These changes were
imported from branch 3 and have been added to the following directory
thus solving the error.