Adds a resolver folder
A resolver folder is added to increase the modularity in the server
code. However, it can not be exported properly onto the index.ts file
to then onboard it. A revision is to be done soon to help with this.
Once it is working, it should allow for better code reusability and
readability.