Overhaul resolvers
This commit changes the resolver functioning. Previously, the resolvers
were not providing proper queries due to improper import of 'Context'.
This has now been changed in the context file. However, this might cause
issues with testing.

Furthermore, three resolvers; userResolvers, businessResolvers,
commentResolvers have been modularised. The rest have to be done as well
More needs to be discussed on how this would affect the testing.