Add rest of the Resolvers
This commit adds rest of the resolvers to the resolver folder for
modularization. Furthermore, unused imports that were present in the
previous commit have been removed. On another note, the mercurius
code generated file has been moved to a build folder to avoid code bulk.
The last issue is to decide wether to remove the resolver file or not.
Further testing needs to be done on the new resolver system.